backend/requirements.txt
Django==4.2.16
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.4.0
psycopg2-binary==2.9.9
python-decouple==3.8
dj-database-url==2.2.0
google-generativeai==0.8.3
openai==1.51.0
anthropic==0.34.2
PyMuPDF==1.24.10
python-docx==1.1.2
Pillow==10.4.0
gunicorn==22.0.0
whitenoise==6.7.0
celery==5.4.0
redis==5.1.1
